Create presentation module and refactor app module
