+++ app/src/main/java/com/sngular/sponsular/SponsularApp.kt
@@ -1,0 +1,6 @@
+package com.sngular.sponsular
+
+import android.app.Application
+
+class SponsularApp : Application() {
+}
