+++ presentation/src/main/java/com/sngular/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.sngular.sponsular
+package com.sngular.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.sngular.sponsular.ui.theme.SponsularTheme
+import com.sngular.presentation.ui.theme.SponsularTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
