+++ presentation/src/main/java/com/sngular/presentation/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.sngular.sponsular.ui.theme
+package com.sngular.presentation.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
