+++ presentation/src/main/java/com/sngular/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.sngular.sponsular.ui.theme
+package com.sngular.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
